test : add test over the death save mechanic + make tests silent
